Clinical_Trial_AI_Assistant-main/Clinical_Trial_AI_Assistant-main/Clinical_2/requirements.txt
streamlit
langchain
langchain-groq
sqlalchemy
mysql-connector-python
python-dotenv
pandas
